uno-main/server/requirements.txt
flask==2.2.3
werkzeug==2.2.3
Flask-SocketIO==5.3.3
Flask-Cors==3.0.10
gunicorn==20.1.0
python-dotenv==1.0.1
gevent==23.9.1
gevent-websocket==0.10.1